+++ android/src/main/kotlin/com/simform/audio_waveforms/AudioPlayer.kt
@@ -1,129 +1,150 @@
 package com.simform.audio_waveforms
 
 import android.content.Context
-import android.os.Build
 import android.os.Handler
 import android.os.Looper
-import androidx.annotation.RequiresApi
 import com.google.android.exoplayer2.ExoPlayer
 import com.google.android.exoplayer2.MediaItem
 import com.google.android.exoplayer2.Player
 import io.flutter.plugin.common.MethodChannel
 import java.lang.Exception
 
-class AudioPlayer(context: Context, channel: MethodChannel) {
-    private val LOG_TAG = "AudioWaveforms"
+class AudioPlayer(context: Context, channel: MethodChannel, playerKey: String) {
     private var handler: Handler = Handler(Looper.getMainLooper())
-    private var runnable = mutableMapOf<String, Runnable?>()
+    private var runnable: Runnable? = null
     private var methodChannel = channel
     private var appContext = context
-    private var players = mutableMapOf<String, ExoPlayer?>()
-    private var playerListeners = mutableMapOf<String, Player.Listener?>()
-    private var preparedPlayers = mutableMapOf<String, Boolean>()
-    private var seekToStart = true
+    private var player: ExoPlayer? = null
+    private var playerListener: Player.Listener? = null
+    private var isPlayerPrepared: Boolean = false
+    private var finishMode = FinishMode.Stop
+    private var key = playerKey
 
-    @RequiresApi(Build.VERSION_CODES.LOLLIPOP)
     fun preparePlayer(
         result: MethodChannel.Result,
         path: String?,
-        volume: Float?,
-        key: String?
+        volume: Float?
     ) {
 
         //TODO: meta data of song
-        if (key != null && path != null) {
+        if (path != null) {
             val mediaItem = MediaItem.fromUri(path)
-            players[key] = ExoPlayer.Builder(appContext).build()
-            players[key]?.addMediaItem(mediaItem)
-            players[key]?.prepare()
-            playerListeners[key] = object : Player.Listener {
+            player = ExoPlayer.Builder(appContext).build()
+            player?.addMediaItem(mediaItem)
+            player?.prepare()
+            playerListener = object : Player.Listener {
                 override fun onPlayerStateChanged(isReady: Boolean, state: Int) {
-                    if (preparedPlayers[key] == false || preparedPlayers[key] == null) {
+                    if (!isPlayerPrepared) {
                         if (state == Player.STATE_READY) {
-                            players[key]?.volume = volume ?: 1F
-                            preparedPlayers[key] = true
+                            player?.volume = volume ?: 1F
+                            isPlayerPrepared = true
                             result.success(true)
                         }
                     }
-                    if (state == Player.STATE_ENDED && seekToStart) {
-                        players[key]?.seekTo(0)
-                        players[key]?.pause()
+                    if (state == Player.STATE_ENDED) {
+                        val args: MutableMap<String, Any?> = HashMap()
+                        when (finishMode) {
+                            FinishMode.Loop -> {
+                                player?.seekTo(0)
+                                player?.play()
+                                args[Constants.finishType] = 0
+                            }
+                            FinishMode.Pause -> {
+                                player?.seekTo(0)
+                                player?.playWhenReady = false
+                                stopListening()
+                                args[Constants.finishType] = 1
+                            }
+                            else -> {
+                                player?.stop()
+                                player?.release()
+                                player = null
+                                stopListening()
+                                args[Constants.finishType] = 2
+                            }
+                        }
+                        args[Constants.playerKey] = key
+                        methodChannel.invokeMethod(
+                            Constants.onDidFinishPlayingAudio,
+                            args
+                        )
                     }
                 }
             }
-            players[key]?.addListener(playerListeners[key]!!)
+            player?.addListener(playerListener!!)
         } else {
-            result.error(LOG_TAG, "path to audio file or unique key can't be null", "")
+            result.error(Constants.LOG_TAG, "path to audio file or unique key can't be null", "")
         }
     }
 
-    @RequiresApi(Build.VERSION_CODES.O)
-    fun seekToPosition(result: MethodChannel.Result, progress: Long?, key: String?) {
-        if (progress != null && key != null) {
-            players[key]?.seekTo(progress)
+    fun seekToPosition(result: MethodChannel.Result, progress: Long?) {
+        if (progress != null) {
+            player?.seekTo(progress)
             result.success(true)
         } else {
             result.success(false)
         }
     }
 
-    fun start(result: MethodChannel.Result, seekToStart: Boolean, key: String?) {
+    fun start(result: MethodChannel.Result, finishMode: Int?) {
         try {
-            this.seekToStart = seekToStart
-            players[key]?.play()
+            if (finishMode != null && finishMode == 0) {
+                this.finishMode = FinishMode.Loop
+            } else if (finishMode != null && finishMode == 1) {
+                this.finishMode = FinishMode.Pause
+            } else {
+                this.finishMode = FinishMode.Stop
+            }
+            player?.playWhenReady = true
+            player?.play()
             result.success(true)
-            startListening(result, key)
+            startListening(result)
         } catch (e: Exception) {
-            result.error(LOG_TAG, "Can not start the player", e.toString())
+            result.error(Constants.LOG_TAG, "Can not start the player", e.toString())
         }
     }
 
-    fun getDuration(result: MethodChannel.Result, durationType: DurationType, key: String?) {
+    fun getDuration(result: MethodChannel.Result, durationType: DurationType) {
         try {
             if (durationType == DurationType.Current) {
-                val duration = players[key]?.currentPosition
+                val duration = player?.currentPosition
                 result.success(duration)
             } else {
-                val duration = players[key]?.duration
+                val duration = player?.duration
                 result.success(duration)
             }
         } catch (e: Exception) {
-            result.error(LOG_TAG, "Can not get duration", e.toString())
+            result.error(Constants.LOG_TAG, "Can not get duration", e.toString())
         }
     }
 
-    fun stop(result: MethodChannel.Result, key: String?) {
-        if (key != null) {
-            stopListening(key)
-            if (playerListeners[key] != null) {
-                players[key]?.removeListener(playerListeners[key]!!)
-                playerListeners.remove(key)
-            }
-            preparedPlayers.remove(key)
-            players[key]?.stop()
-            players[key]?.release()
-            result.success(true)
+    fun stop(result: MethodChannel.Result) {
+        stopListening()
+        if (playerListener != null) {
+            player?.removeListener(playerListener!!)
         }
+        isPlayerPrepared = false
+        player?.stop()
+        player?.release()
+        result.success(true)
     }
 
 
-    fun pause(result: MethodChannel.Result, key: String?) {
-        if (key != null) {
-            try {
-                stopListening(key)
-                players[key]?.pause()
-                result.success(true)
-            } catch (e: Exception) {
-                result.error(LOG_TAG, "Failed to pause the player", e.toString())
-            }
-        }
-
-    }
-
-    fun setVolume(volume: Float?, result: MethodChannel.Result, key: String?) {
+    fun pause(result: MethodChannel.Result) {
         try {
-            if (volume != null && key != null) {
-                players[key]?.volume = volume
+            stopListening()
+            player?.pause()
+            result.success(true)
+        } catch (e: Exception) {
+            result.error(Constants.LOG_TAG, "Failed to pause the player", e.toString())
+        }
+
+    }
+
+    fun setVolume(volume: Float?, result: MethodChannel.Result) {
+        try {
+            if (volume != null) {
+                player?.volume = volume
                 result.success(true)
             } else {
                 result.success(false)
@@ -133,40 +154,26 @@ class AudioPlayer(context: Context, channel: MethodChannel) {
         }
     }
 
-    private fun startListening(result: MethodChannel.Result, key: String?) {
-        if (key != null) {
-            runnable[key] = object : Runnable {
-                override fun run() {
-                    val currentPosition = players[key]?.currentPosition
-                    if (currentPosition != null) {
-                        val args: MutableMap<String, Any?> = HashMap()
-                        args[Constants.current] = currentPosition
-                        args[Constants.playerKey] = key
-                        methodChannel.invokeMethod(Constants.onCurrentDuration, args)
-                        handler.postDelayed(this, 200)
-                    } else {
-                        result.error(LOG_TAG, "Can't get current Position of player", "")
-                    }
+    private fun startListening(result: MethodChannel.Result) {
+        runnable = object : Runnable {
+            override fun run() {
+                val currentPosition = player?.currentPosition
+                if (currentPosition != null) {
+                    val args: MutableMap<String, Any?> = HashMap()
+                    args[Constants.current] = currentPosition
+                    args[Constants.playerKey] = key
+                    methodChannel.invokeMethod(Constants.onCurrentDuration, args)
+                    handler.postDelayed(this, 200)
+                } else {
+                    result.error(Constants.LOG_TAG, "Can't get current Position of player", "")
                 }
             }
-            handler.post(runnable[key]!!)
         }
+        handler.post(runnable!!)
 
     }
 
-    private fun stopListening(key: String?) {
-        runnable[key]?.let { handler.removeCallbacks(it) }
-    }
-
-    fun stopAllPlayers(result: MethodChannel.Result) {
-        for ((key, _) in players) {
-            players[key]?.stop()
-            players[key] = null
-        }
-        for ((key, _) in runnable) {
-            runnable[key]?.let { handler.removeCallbacks(it) }
-            runnable[key] = null
-        }
-        result.success(true)
+    private fun stopListening() {
+        runnable?.let { handler.removeCallbacks(it) }
     }
 }