+++ android/src/main/kotlin/com/simform/audio_waveforms/AudioWaveformsPlugin.kt
@@ -1,6 +1,7 @@
 package com.simform.audio_waveforms
 
 import android.app.Activity
+import android.content.Context
 import android.media.MediaRecorder
 import android.os.Build
 import android.util.Log
@@ -30,16 +31,16 @@ class AudioWaveformsPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
     private var encoder: Int = 0
     private var outputFormat: Int = 0
     private var sampleRate: Int = 16000
+    private lateinit var applicationContext: Context
 
     //Todo: bitrate
-    private lateinit var audioPlayer: AudioPlayer
+    private var audioPlayers = mutableMapOf<String, AudioPlayer?>()
 
     override fun onAttachedToEngine(@NonNull flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
         channel = MethodChannel(flutterPluginBinding.binaryMessenger, Constants.methodChannelName)
         channel.setMethodCallHandler(this)
         audioRecorder = AudioRecorder()
-        audioPlayer = AudioPlayer(flutterPluginBinding.applicationContext, channel)
-
+        applicationContext = flutterPluginBinding.applicationContext
     }
 
     @RequiresApi(Build.VERSION_CODES.N)
@@ -65,26 +66,51 @@ class AudioWaveformsPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
                 val audioPath = call.argument(Constants.path) as String?
                 val volume = call.argument(Constants.volume) as Double?
                 val key = call.argument(Constants.playerKey) as String?
-                audioPlayer.preparePlayer(result, audioPath, volume?.toFloat(), key)
+                if (key != null) {
+                    initPlayer(key)
+                    audioPlayers[key]?.preparePlayer(result, audioPath, volume?.toFloat())
+                } else {
+                    result.error(Constants.LOG_TAG, "Player key can't be null", "")
+                }
+
             }
             Constants.startPlayer -> {
-                val seekToStart = call.argument(Constants.seekToStart) as Boolean?
+                val finishMode = call.argument(Constants.finishMode) as Int?
                 val key = call.argument(Constants.playerKey) as String?
-                audioPlayer.start(result, seekToStart ?: true, key)
+                if (key != null) {
+                    audioPlayers[key]?.start(
+                        result,
+                        finishMode ?: 2
+                    )
+                } else {
+                    result.error(Constants.LOG_TAG, "Player key can't be null", "")
+                }
             }
             Constants.stopPlayer -> {
                 val key = call.argument(Constants.playerKey) as String?
-                audioPlayer.stop(result, key)
+                if (key != null) {
+                    audioPlayers[key]?.stop(result)
+                } else {
+                    result.error(Constants.LOG_TAG, "Player key can't be null", "")
+                }
             }
             Constants.pausePlayer -> {
                 val key = call.argument(Constants.playerKey) as String?
-                audioPlayer.pause(result, key)
+                if (key != null) {
+                    audioPlayers[key]?.pause(result)
+                } else {
+                    result.error(Constants.LOG_TAG, "Player key can't be null", "")
+                }
             }
             Constants.seekTo -> {
                 if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
                     val progress = call.argument(Constants.progress) as Int?
                     val key = call.argument(Constants.playerKey) as String?
-                    audioPlayer.seekToPosition(result, progress?.toLong(), key)
+                    if (key != null) {
+                        audioPlayers[key]?.seekToPosition(result, progress?.toLong())
+                    } else {
+                        result.error(Constants.LOG_TAG, "Player key can't be null", "")
+                    }
                 } else {
                     Log.e(
                         Constants.LOG_TAG,
@@ -95,15 +121,29 @@ class AudioWaveformsPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
             Constants.setVolume -> {
                 val volume = call.argument(Constants.volume) as Float?
                 val key = call.argument(Constants.playerKey) as String?
-                audioPlayer.setVolume(volume, result, key)
+                if (key != null) {
+                    audioPlayers[key]?.setVolume(volume, result)
+                } else {
+                    result.error(Constants.LOG_TAG, "Player key can't be null", "")
+                }
             }
             Constants.getDuration -> {
                 val type =
                     if ((call.argument(Constants.durationType) as Int?) == 0) DurationType.Current else DurationType.Max
                 val key = call.argument(Constants.playerKey) as String?
-                audioPlayer.getDuration(result, type, key)
+                if (key != null) {
+                    audioPlayers[key]?.getDuration(result, type)
+                } else {
+                    result.error(Constants.LOG_TAG, "Player key can't be null", "")
+                }
             }
-            Constants.stopAllPlayers -> audioPlayer.stopAllPlayers(result)
+            Constants.stopAllPlayers -> {
+                for ((key, _) in audioPlayers) {
+                    audioPlayers[key]?.stop(result)
+                    audioPlayers[key] = null
+                }
+                result.success(true)
+            }
             else -> result.notImplemented()
         }
     }
@@ -149,6 +189,14 @@ class AudioWaveformsPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
                 sampleRate
             )
         }
+    }
+
+    private fun initPlayer(playerKey: String) {
+        if (audioPlayers[playerKey] == null) {
+            val newPlayer = AudioPlayer(applicationContext, channel, playerKey)
+            audioPlayers[playerKey] = newPlayer
+        }
+        return
     }
 
     override fun onDetachedFromEngine(@NonNull binding: FlutterPlugin.FlutterPluginBinding) {