+++ lib/src/base/player_indentifier.dart
@@ -1,0 +1,10 @@
+///This class is to identify player associated with any type.
+class PlayerIdentifier<T> {
+  ///An unique key associated with player.
+  String playerKey;
+
+  ///A generic type which is associated to player
+  T type;
+
+  PlayerIdentifier(this.playerKey, this.type);
+}