+++ lib/src/controllers/player_controller.dart
@@ -3,9 +3,13 @@ import 'dart:io';
 import 'dart:typed_data';
 
 import 'package:audio_waveforms/audio_waveforms.dart';
-import 'package:audio_waveforms/src/base/audio_waveforms_interface.dart';
+import 'package:audio_waveforms/src/base/constants.dart';
 import 'package:audio_waveforms/src/base/platform_streams.dart';
+import 'package:audio_waveforms/src/base/player_indentifier.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+
+part '../base/audio_waveforms_interface.dart';
 
 class PlayerController extends ChangeNotifier {
   Uint8List? _bufferData;
@@ -18,6 +22,12 @@ class PlayerController extends ChangeNotifier {
   ///provides current state of the player
   PlayerState get playerState => _playerState;
 
+  void setPlayerState(PlayerState state) {
+    _playerState = state;
+    PlatformStreams.instance
+        .addPlayerStateEvent(PlayerIdentifier(playerKey, state));
+  }
+
   String? _audioFilePath;
 
   int _maxDuration = -1;
@@ -25,17 +35,26 @@ class PlayerController extends ChangeNotifier {
   ///provides [max] duration of currenly provided audio file.
   int get maxDuration => _maxDuration;
 
-  bool _seekToStart = true;
-
   final UniqueKey _playerKey = UniqueKey();
 
+  ///An unique key string associated with [this] player only
   String get playerKey => _playerKey.toString();
+
+  ///A stream to get current state of the player. This stream
+  ///will emit event whenever there is change in the playerState.
+  Stream<PlayerState> get onPlayerStateChanged =>
+      PlatformStreams.instance.onPlayerStateChanged.filter(playerKey);
+
+  ///A Stream to get current duration. This stream will emit
+  ///duration in milliseconds every 200 milliseconds.
+  Stream<int> get onCurrentDurationChanged =>
+      PlatformStreams.instance.onDurationChanged.filter(playerKey);
 
   PlayerController() {
     if (!PlatformStreams.instance.isInitialised) {
       PlatformStreams.instance.init();
     }
-    AudioWaveformsInterface.instance.setMethodCallHandler();
+    PlatformStreams.instance.playerControllerFactory.addAll({playerKey: this});
   }
 
   ///Reads bytes from audio file
@@ -46,7 +65,7 @@ class PlayerController extends ChangeNotifier {
       var bytes = await file.readAsBytes();
       _bufferData = bytes;
       if (_bufferData != null) {
-        _playerState = PlayerState.readingComplete;
+        setPlayerState(PlayerState.readingComplete);
       } else {
         throw "Can't read given audio file";
       }
@@ -61,7 +80,8 @@ class PlayerController extends ChangeNotifier {
   ///It first reads bytes from audio file so as soon as it completes
   /// it prepares audio player.
   ///
-  ///[playerState] has to be PlayerState.readingComplete otherwise throws [Exception].
+  ///[playerState] has to be PlayerState.readingComplete when preparing player
+  /// otherwise throws [Exception].
   ///
   ///This behavior is set to ensure that player is only re-initialised for new audio file.
   Future<void> preparePlayer(String path, [double? volume]) async {
@@ -72,7 +92,7 @@ class PlayerController extends ChangeNotifier {
           .preparePlayer(path, _playerKey.toString(), volume);
       if (isPrepared) {
         _maxDuration = await getDuration();
-        _playerState = PlayerState.initialized;
+        setPlayerState(PlayerState.initialized);
       }
       notifyListeners();
     } else {
@@ -82,17 +102,16 @@ class PlayerController extends ChangeNotifier {
 
   ///Use this function to start the player to play/resume the audio.
   ///
-  ///When playing audio is finished player will be seeked to [start]. To change
-  ///this behaviour pass false for [seekToStart] parameter and player position will
-  ///stay at last
-  Future<void> startPlayer([bool? seekToStart]) async {
+  ///When playing audio is finished [player] will be [stopped] and [disposed]
+  ///by default. To change this behavior use [FinishMode] enum.
+  ///See FinishMode documentation for more info.
+  Future<void> startPlayer({FinishMode finishMode = FinishMode.stop}) async {
     if (_playerState == PlayerState.initialized ||
         _playerState == PlayerState.paused) {
-      _seekToStart = seekToStart ?? true;
       final isStarted = await AudioWaveformsInterface.instance
-          .startPlayer(_playerKey.toString(), _seekToStart);
+          .startPlayer(_playerKey.toString(), finishMode);
       if (isStarted) {
-        _playerState = PlayerState.playing;
+        setPlayerState(PlayerState.playing);
       } else {
         throw "Failed to start player";
       }
@@ -100,12 +119,12 @@ class PlayerController extends ChangeNotifier {
     notifyListeners();
   }
 
-  ///Use this to pause the playing audio
+  ///Use this to pause the playing audio.
   Future<void> pausePlayer() async {
     final isPaused = await AudioWaveformsInterface.instance
         .pausePlayer(_playerKey.toString());
     if (isPaused) {
-      _playerState = PlayerState.paused;
+      setPlayerState(PlayerState.paused);
     }
     notifyListeners();
   }
@@ -115,7 +134,7 @@ class PlayerController extends ChangeNotifier {
     final isStopped = await AudioWaveformsInterface.instance
         .stopPlayer(_playerKey.toString());
     if (isStopped) {
-      _playerState = PlayerState.stopped;
+      setPlayerState(PlayerState.stopped);
     }
     notifyListeners();
   }
@@ -143,7 +162,7 @@ class PlayerController extends ChangeNotifier {
 
   ///Moves the media to specified time position. pass progress parameter in milliseconds.
   ///
-  /// Minimum Android O is required to use this funtion otherwise nothing happens.
+  /// Minimum Android [O] is required to use this funtion otherwise nothing happens.
   Future<void> seekTo(int progress) async {
     if (progress < 0) return;
     if (_playerState == PlayerState.playing) {
@@ -155,15 +174,17 @@ class PlayerController extends ChangeNotifier {
   ///Calling this will stop the player, it will dispose player in native and
   ///also dispose the the controller.
   ///
-  ///As there is common stream for every players stream
+  ///As there is common stream for every players, stream
   /// will be still active without any events.
   /// To dispose it call [stopAllPlayers].
   ///
-  /// Only calling disposeFunc() will not stop the stream
-  /// calling [stopAllPlayers] is still [required]
-  void disposeFunc() async {
+  /// Only calling dispose() will not stop the stream
+  /// calling [stopAllPlayers] is still [required].
+  @override
+  void dispose() async {
     if (playerState != PlayerState.stopped) await stopPlayer();
-    dispose();
+    PlatformStreams.instance.playerControllerFactory.remove(this);
+    super.dispose();
   }
 
   ///This method is to free all players [resources] all at once.
@@ -171,12 +192,18 @@ class PlayerController extends ChangeNotifier {
   /// This method is required to call only [once] from any
   /// one of the [PlayerController]s.
   ///
-  /// Note -:this method will close stream and free  resources taken by players but
-  /// it will only dispose controller who's stopAllPlayers method was called
-  /// calling [disposeFunc] is still required for every other [PlayerController]
+  /// Note -:this method will close stream and free resources taken by players
+  /// but it will only dispose controller who's stopAllPlayers method was called
+  /// calling [dispose] is still required for every other [PlayerController]
   void stopAllPlayers() async {
     PlatformStreams.instance.dispose();
     await AudioWaveformsInterface.instance.stopAllPlayers();
-    dispose();
+    PlatformStreams.instance.playerControllerFactory.remove(this);
+    super.dispose();
+  }
+
+  @override
+  bool operator ==(Object other) {
+    return other is PlayerController && other.playerKey == playerKey;
   }
 }